aiogramarch/cli/templates/project/{{ cookiecutter.project_name }}/requirements.txt
aiogram==3.0.0b9
loguru==0.6.0
aerich==0.7.2
tortoise-orm==0.19.3
pydantic-settings==2.0.2
python-dotenv
